Norme done √

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: rolaforg <rolaforg@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/11/11 15:56:46 by rolaforg     #+#   ##    ##    #+#       */
-/*   Updated: 2020/02/03 19:31:58 by rolaforg    ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/02/04 15:19:49 by rolaforg    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -22,11 +22,11 @@
 # include <fcntl.h>
 
 int		get_next_line(int fd, char **line);
-char	*ft_substrdup(char *s, unsigned int start, size_t len);
-char	*ft_substr(char *s, unsigned int start, size_t len);
+char	*ft_substrdup(char *s, unsigned long start, size_t len);
+char	*ft_substr(char *s, unsigned long start, size_t len);
 void	*ft_calloc(size_t count, size_t size);
 char	*ft_strjoin(char const *s1, char const *s2);
-char	*ft_memchr(char *s, int c, size_t n);
+char	*ft_memchr(char *s, long c, size_t n);
 size_t	ft_strlen(char *str);
 size_t	ft_linelen(char *str);
